user-service/requirements.txt
Babel==2.11.0
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
flask-babel==3.0.0
Flask-Bcrypt==1.0.1
Flask-Table==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyMySQL==1.0.2
pytz==2022.7.1
Werkzeug==2.2.2
mysql-connector-python==8.0.31
mysql-connector-python-rf==2.2.2
